Delete the old load/stress test code
fixes HNT-5021
