+++ .vscode/launch.json
@@ -71,30 +71,5 @@
       "args": ["${file}", "--config", "${workspaceFolder}/core/sdk/jest.config.ts", "-i", "--no-cache", "--forceExit"],
       "console": "integratedTerminal",
     },
-    {
-      "name": "Stress Jest: current file in 'core/sdk/' (single)",
-      "type": "node",
-      "request": "launch",
-      "env": { "NODE_ENV": "development", "NODE_TLS_REJECT_UNAUTHORIZED": "0", "RIVER_ENV": "local_single_ne", "DEBUG": "csb:test:stress*", "DEBUG_DEPTH":"10" },
-      "program": "${workspaceFolder}/node_modules/.bin/jest",
-      "runtimeArgs": ["--experimental-vm-modules", "--experimental-wasm-modules"],
-      "args": ["${file}", "--config", "${workspaceFolder}/core/sdk/jest.config.ts", "-i", "--no-cache", "--logHeapUsage", "--runInBand", "--testTimeout=7200000"],
-      "console": "integratedTerminal",
-    },
-    {
-      "name": "Casablanca Core Jest Tests Jsdom",
-      "type": "node",
-      "request": "launch",
-      "runtimeArgs": [
-        "--inspect-brk",
-        "${workspaceRoot}/node_modules/jest/bin/jest.js",
-        "--runInBand",
-        "--config",
-        "${workspaceRoot}/core/core/jest.config.jsdom.ts",
-        "${fileBasenameNoExtension}",
-      ],
-      "console": "integratedTerminal",
-      "internalConsoleOptions": "neverOpen"
-    },
   ]
 }
