--- core/sdk/src/load/zk-single-kafka-single.yml
@@ -1,85 +1,0 @@
-version: '3'
-
-services:
-    stress_test_zoo:
-        image: confluentinc/cp-zookeeper:latest
-        hostname: stress_test_zoo
-        container_name: stress_test_zoo
-        ports:
-            - '2181:2181'
-        environment:
-            ZOOKEEPER_CLIENT_PORT: 2181
-            ZOOKEEPER_SERVER_ID: 1
-            ZOOKEEPER_SERVERS: stress_test_zoo:2888:3888
-            ZOOKEPER_TICK_TIME: 2000
-
-    stress_test_kafka:
-        image: confluentinc/cp-kafka:latest
-        hostname: stress_test_kafka
-        container_name: stress_test_kafka
-        ports:
-            - '9092:9092'
-            - '29092:29092'
-            - '9999:9999'
-            - '19092:19092'
-            - '29093:29093'
-        environment:
-            KAFKA_LISTENERS: PLAINTEXT://:9092, INTERNAL://:19092,EXTERNAL://:29092,DOCKER://:29093
-            KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://stress_test_kafka:9092, INTERNAL://stress_test_kafka:19092,EXTERNAL://${DOCKER_HOST_IP:-127.0.0.1}:29092,DOCKER://host.docker.internal:29093
-            KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: PLAINTEXT:PLAINTEXT, INTERNAL:PLAINTEXT,EXTERNAL:PLAINTEXT,DOCKER:PLAINTEXT
-            KAFKA_INTER_BROKER_LISTENER_NAME: INTERNAL
-            KAFKA_ZOOKEEPER_CONNECT: 'stress_test_zoo:2181'
-            KAFKA_BROKER_ID: 1
-            KAFKA_LOG4J_LOGGERS: 'kafka.controller=INFO,kafka.producer.async.DefaultEventHandler=INFO,state.change.logger=INFO'
-            KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
-            KAFKA_TRANSACTION_STATE_LOG_REPLICATION_FACTOR: 1
-            KAFKA_TRANSACTION_STATE_LOG_MIN_ISR: 1
-            KAFKA_JMX_PORT: 9999
-            KAFKA_JMX_HOSTNAME: ${DOCKER_HOST_IP:-127.0.0.1}
-            KAFKA_AUTHORIZER_CLASS_NAME: kafka.security.authorizer.AclAuthorizer
-            KAFKA_ALLOW_EVERYONE_IF_NO_ACL_FOUND: 'true'
-        depends_on:
-            - stress_test_zoo
-    stress_test_schema_registry:
-        image: confluentinc/cp-schema-registry:latest
-        hostname: stress_test_schema_registry
-        ports:
-            - '8081:8081'
-        depends_on:
-            - stress_test_kafka
-            - stress_test_zoo
-        environment:
-            SCHEMA_REGISTRY_KAFKASTORE_BOOTSTRAP_SERVERS: PLAINTEXT://stress_test_kafka:9092
-            SCHEMA_REGISTRY_KAFKASTORE_SECURITY_PROTOCOL: PLAINTEXT
-            SCHEMA_REGISTRY_KAFKASTORE_CONNECTION_URL: 'stress_test_zoo:2181'
-            SCHEMA_REGISTRY_HOST_NAME: stress_test_schema_registry
-            SCHEMA_REGISTRY_LISTENERS: http://0.0.0.0:8081
-            SCHEMA_REGISTRY_DEBUG: 'true'
-    stress_test_kafka_rest_proxy:
-        image: confluentinc/cp-kafka-rest:latest
-        hostname: stress_test_kafka_rest_proxy
-        container_name: stress_test_kafka_rest_proxy
-        depends_on:
-            - stress_test_kafka
-            - stress_test_schema_registry
-            - stress_test_zoo
-        ports:
-            - '8082:8082'
-        environment:
-            KAFKA_REST_HOST_NAME: kafka-rest-proxy
-            KAFKA_REST_BOOTSTRAP_SERVERS: stress_test_kafka:9092
-            KAFKA_REST_LISTENERS: http://0.0.0.0:8082
-            KAFKA_REST_SCHEMA_REGISTRY_URL: http://stress_test_kafka_schema_registry:8081
-            KAFKA_REST_ACCESS_CONTROL_ALLOW_ORIGIN: '*'
-            KAFKA_REST_ACCESS_CONTROL_ALLOW_METHODS: 'GET,POST,PUT,DELETE'
-            KAFKA_REST_ACCESS_CONTROL_ALLOW_HEADERS: 'origin,content-type,accept,authorization'
-    stress_test_redis:
-        image: redis:latest
-        hostname: stress_test_redis
-        container_name: stress_test_redis
-        restart: always
-        ports:
-            - '6379:6379'
-        environment:
-            - REDIS_PORT=6379
-            - REDIS_DATABASES=1
